refactor and update readme documentation with events
