+++ demo/index.html
@@ -52,8 +52,8 @@
 							<d2l-labs-d2l-grade-result
 								href="https://648df3aa-e573-4c44-b34b-856c0523e9d5.activities.api.proddev.d2l/activities/6606_3000_2/usages/6609/users/178/grade"
 								token="eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6Ijg5ZmU2ZDcwLTMxZmEtNGM0YS05NGNmLTA4OTY1ZWZmNmVjMSJ9.eyJzdWIiOiIxNzciLCJ0ZW5hbnRpZCI6IjY0OGRmM2FhLWU1NzMtNGM0NC1iMzRiLTg1NmMwNTIzZTlkNSIsInNjb3BlIjoiKjoqOioiLCJqdGkiOiI2ZDU0NzNlOC0zNzFlLTQ1YmEtOTBjMS1iMmUxMDMxZTNiOTYiLCJpc3MiOiJodHRwczovL2FwaS5icmlnaHRzcGFjZS5jb20vYXV0aCIsImF1ZCI6Imh0dHBzOi8vYXBpLmJyaWdodHNwYWNlLmNvbS9hdXRoL3Rva2VuIiwiZXhwIjoxNTg2MjAyNzg5LCJuYmYiOjE1ODYxOTkxODl9.BVyB8EAk8UAfuwDkLNeqAdoc5knZdtOLzvdU4Hv5rctBsCzD1PZ72J484MolsLo7jdnkyNfO8o7sOZIdl3bgxLuC3qNfF_ho-qmCIYj3A3xIKx5uqj0tCoKaNk3t88MX930ZWUq16E3ZaerI1ViZkUWo-psncRMqMgiJ4-oToNOuD2xRBt6SMrPVgRcyCcBQOaRLEU5uIAwr8XVGNSWMl8TD9iLIhBqeoGtCUCI3uPQ5KTJODRTG1aNGDFZfYUo2BpLLkwTDd0dc2GPOpgUl2ppS75ba2GKVoOfnSNi9tNPelT8r9uHd3GLpM0Vd578oN5QlihcLXIbwP9GVUXK9Eg"
-								@d2l-grade-result-intitialized-success=${this.onInitializedSuccess}
-								@d2l-grade-result-intitialized-error=${this.onInitializedError}
+								@d2l-grade-result-initialized-success=${this.onInitializedSuccess}
+								@d2l-grade-result-initialized-error=${this.onInitializedError}
 								@d2l-grade-result-grade-updated-success=${this.onGradeUpdatedSuccess}
 								@d2l-grade-result-grade-updated-error=${this.onGradeUpdatedError}
 							></d2l-labs-d2l-grade-result>
