+++ src/components/d2l-grade-result-icon-button.js
