+++ src/components/d2l-grade-result-numeric-score.js
