+++ src/components/locale.js
@@ -3,7 +3,7 @@ export default async function getLocalizationTranslations(langs) {
 		let translations;
 		switch (lang) {
 			case 'en':
-				translations = await import('../locales/en.js');
+				translations = await import('../../locales/en.js');
 				break;
 		}
