+++ test/perceptual/d2l-grade-result-read-only.visual-diff.html
@@ -5,7 +5,7 @@
 		<script type="module">
 			import '@brightspace-ui/core/components/typography/typography.js';
 			import '../../d2l-grade-result.js';
-			import '../../components/d2l-grade-result-presentational.js';
+			import '../../src/components/d2l-grade-result-presentational.js';
 		</script>
 		<title>d2l-labs-d2l-grade-result visual-diff tests</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
